[refact] TEST 시나리오별 파일구조 변경
