+++ CRA_Project_Tester/CRA_Project_Tester.vcxproj
@@ -130,14 +130,12 @@
   <ItemGroup>
     <ClCompile Include="Logger.cpp" />
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="featureTEST.cpp" />
-    <ClCompile Include="test_ara.cpp" />
-    <ClCompile Include="test_run.cpp">
+    <ClCompile Include="testcode_scenario.cpp">
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">   %(AdditionalOptions)</AdditionalOptions>
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Release|x64'">   %(AdditionalOptions)</AdditionalOptions>
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Release|x64'">false</ExcludedFromBuild>
     </ClCompile>
-    <ClCompile Include="test_suyong.cpp" />
+    <ClCompile Include="testcode_feature.cpp" />
     <ClCompile Include="testScript.cpp">
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">/TP %(AdditionalOptions)</AdditionalOptions>
     </ClCompile>
