--- CRA_Project_Tester/featureTEST.cpp
@@ -1,59 +1,0 @@
-#include "gmock/gmock.h"
-#include <string>
-#include <iostream>
-#include "testShell.h"
-
-using namespace testing;
-using namespace std;
-
-
-
-class MockReadDriver : public Read {
-public:
-    MOCK_METHOD(string, read, (string cmd1), (override));
-
-    MOCK_METHOD(void, run, (string, string), (override));
-
-};
-class MockFullRead : public FullRead {
-public:
-    MOCK_METHOD(string, read, (string address), (override));
-};
-
-TEST(SSDTEST, ReadNone) {
-
-    MockReadDriver mockDriver;
-    EXPECT_CALL(mockDriver, read("3")).WillRepeatedly(Return("helloWorld"));
-
-    EXPECT_EQ(mockDriver.read("3"), "helloWorld");
-}
-
-TEST(SSDTEST, LBAinvalid) {
-
-    MockReadDriver mockDriver;
-    EXPECT_NO_THROW({
-    bool result = mockDriver.checkLBA("12");
-    EXPECT_TRUE(result);
-    });
-}
-
-TEST(SSDTEST, LBAinvalid2) {
-
-    MockReadDriver mockDriver;
-    EXPECT_THROW(mockDriver.checkLBA("123"), std::invalid_argument);
-}
-
-TEST(SSDTEST, LBAinvalid3) {
-    MockReadDriver mockDriver;
-    EXPECT_THROW(mockDriver.checkLBA("1A"), std::invalid_argument);
-}
-
-TEST(FullReadTest, CallsRead100Times) {
-    MockFullRead mock;
-
-    EXPECT_CALL(mock, read(_))  
-        .Times(Exactly(100))
-        .WillRepeatedly(Return("mocked_data"));
-
-    mock.run();
-}
